+++ src/lib/components/SmartList/index.jsx
@@ -1,20 +1,25 @@
 import PropTypes from 'prop-types'
 
+import object from '~su/utilities/object'
+
 import Card from '../Card'
 import List from '../List'
 
-const SmartList = ({ title, onChange, pagination, ...listProps }) => {
+const SmartList = ({ title, onChange, pagination, extraDataDisplay, ...listProps }) => {
   const applyPagination = (page) => {
     return onChange({ current: page }, {}, undefined, { action: 'paginate' })
   }
 
-  pagination = {
-    ...pagination,
-    ...(onChange && { onChange: applyPagination })
-  }
+  pagination = object.isEmpty(pagination)
+    ? false
+    : {
+        ...pagination,
+        ...(onChange && { onChange: applyPagination })
+      }
 
   return (
     <Card title={title} headStyle={{ fontWeight: 'normal', border: 'none' }}>
+      {extraDataDisplay}
       <List pagination={pagination} {...listProps} />
     </Card>
   )
@@ -23,7 +28,8 @@ const SmartList = ({ title, onChange, pagination, ...listProps }) => {
 SmartList.propTypes = {
   title: PropTypes.element,
   onChange: PropTypes.func,
-  pagination: PropTypes.oneOfType([PropTypes.bool, PropTypes.object])
+  pagination: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
+  extraDataDisplay: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)])
 }
 
 export default SmartList