+++ src/lib/components/TagsList.jsx
@@ -1,14 +1,20 @@
+import PropTypes from 'prop-types'
+
 import number from '~su/utilities/number'
 
 import Tag from './Tag'
 import Space from './Space'
 
-const TagsList = ({ tags, ...space }) => {
+const TagsList = ({ tags, tagSize = 'default', ...space }) => {
   const renderTag = (tag, index) => {
     const key = `${tag}_${index}_${number.randomInt()}`,
       value = tag
 
-    return <Tag key={key}>{value}</Tag>
+    return (
+      <Tag key={key} size={tagSize}>
+        {value}
+      </Tag>
+    )
   }
 
   return tags ? (
@@ -16,6 +22,11 @@ const TagsList = ({ tags, ...space }) => {
       {tags.map(renderTag)}
     </Space>
   ) : null
+}
+
+TagsList.propTypes = {
+  tags: PropTypes.arrayOf(PropTypes.string),
+  size: PropTypes.oneOf(['large', 'default', null])
 }
 
 export default TagsList