+++ src/lib/components/index.js
@@ -12,6 +12,7 @@ import ExpandableEditableTable from './EditableTable/Expandable'
 import ErrorsList from './ErrorsList'
 import InlineEditableTable from './EditableTable/Inline'
 import Form from './Form'
+import Flex from './Flex'
 import GlobalAlert from './GlobalAlert'
 import Field from './Form/Field'
 import FieldsList from './Form/FieldsList'
@@ -41,6 +42,7 @@ import IconButton from './IconButton'
 import Input from './Input'
 import Layout from './Layout'
 import List from './List'
+import LoadingBlock from './LoadingBlock'
 import MarkdownEditor from './MarkdownEditor'
 import Menu from './Menu'
 import Modal from './Modal'
@@ -57,6 +59,7 @@ import Spacer from './Spacer'
 import StateBadge from './StateBadge'
 import StatesHistory from './StatesHistory'
 import StateTag from './StateTag'
+import Steps from './Steps'
 import StyledIcon from './StyledIcon'
 import Switch from './Switch'
 import Table from './Table'
@@ -64,6 +67,7 @@ import Tabs from './Tabs'
 import Tag from './Tag'
 import TagsList from './TagsList'
 import TextWithCountBadge from './TextWithCountBadge'
+import Timeline from './Timeline'
 import TimeTaken from './TimeTaken'
 import ToggledSelect from './ToggledSelect'
 import Tooltip from './Tooltip'
@@ -89,6 +93,7 @@ export {
   ErrorsList,
   InlineEditableTable,
   Form,
+  Flex,
   GlobalAlert,
   Field,
   FieldsList,
@@ -118,6 +123,7 @@ export {
   Input,
   Layout,
   List,
+  LoadingBlock,
   MarkdownEditor,
   Menu,
   Modal,
@@ -134,6 +140,7 @@ export {
   StateBadge,
   StatesHistory,
   StateTag,
+  Steps,
   StyledIcon,
   Switch,
   Table,
@@ -141,6 +148,7 @@ export {
   Tag,
   TagsList,
   TextWithCountBadge,
+  Timeline,
   TimeTaken,
   ToggledSelect,
   Tooltip,